Finised off the encryption client side. Fixed the dm feature. Fixed small issue with server. Started integration with textual.
Started implementation with textual, run the program and it works!
tested the dm feature and it works! tested the encryption and it works!
really cool, very satisfying. now i need to work on the front end with
textual - lets see how much of a front end dev i can be...